requirements.txt
flask
gspread
oauth2client
python-dotenv
google-generativeai
